Release integration tests jar in maven central (#4807)
* release tests jar for project dirigible-tests

* make tests public

* explicitly include src/test/resources

* add CommonTestSuite

* ensure test jars are deployed in maven central

* remove DirigibleJavaScriptComponentIT.java since it duplicates another test

* rollback handler.mjs

* return DirigibleJavaScriptComponentIT.java

* support copy from jars as well

* test-routes.camel

* restructure dependencies

* dependency refactoring

+++ tests/tests-framework/src/main/java/org/eclipse/dirigible/tests/util/ProjectUtil.java
@@ -19,19 +19,22 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
-import java.io.File;
-import java.io.IOException;
+import java.io.*;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.Enumeration;
 import java.util.Map;
+import java.util.jar.JarEntry;
+import java.util.jar.JarFile;
 
 @Component
 public class ProjectUtil {
 
+    public static final String JAR_PATH = ".jar!";
+    public static final String FILE_PREFIX = "file:/";
     private static final Logger LOGGER = LoggerFactory.getLogger(ProjectUtil.class);
-
     private final IRepository repository;
 
     ProjectUtil(IRepository repository) {
@@ -66,15 +69,21 @@ public class ProjectUtil {
         copyResourceFolder(resourcesFolder, destinationDirPath, placeholders);
     }
 
-    private void copyResourceFolder(String resourcesFolder, String destinationDirPath, Map<String, String> placeholders) {
+    public void copyResourceFolder(String resourcesFolder, String destinationDirPath, Map<String, String> placeholders) {
         String sourceDirPath = getResourcePath(resourcesFolder);
         File sourceDir = new File(sourceDirPath);
-
         File destinationDir = new File(destinationDirPath);
 
         LOGGER.debug("Copy folder [{}] to folder [{}]", sourceDir, destinationDir);
+
         try {
-            FileUtils.copyDirectory(sourceDir, destinationDir);
+            if (sourceDirPath.contains(JAR_PATH)) {
+                copyFromJar(sourceDirPath, destinationDir);
+            } else if (sourceDir.exists()) {
+                FileUtils.copyDirectory(sourceDir, destinationDir);
+            } else {
+                throw new IllegalStateException("Source is neither a valid directory nor a valid JAR resource: " + sourceDirPath);
+            }
         } catch (IOException ex) {
             throw new IllegalStateException("Failed to copy [" + sourceDir + "] to " + destinationDir, ex);
         }
@@ -88,10 +97,56 @@ public class ProjectUtil {
     private String getResourcePath(String path) {
         URL resourceURL = ProjectUtil.class.getClassLoader()
                                            .getResource(path);
-        if (null == resourceURL) {
+        if (resourceURL == null) {
             throw new IllegalStateException("Missing resource with path " + path);
         }
         return resourceURL.getPath();
+    }
+
+    private void copyFromJar(String resourcePath, File destinationDir) throws IOException {
+        int beginIndex = resourcePath.indexOf(FILE_PREFIX) + FILE_PREFIX.length() - 1;
+        int endIndex = resourcePath.indexOf(JAR_PATH) + JAR_PATH.length() - 1;
+        String jarPath = resourcePath.substring(beginIndex, endIndex);
+        String resourcePathInJar = resourcePath.substring(resourcePath.lastIndexOf(JAR_PATH) + JAR_PATH.length() + 1);
+
+        try (JarFile jarFile = new JarFile(jarPath)) {
+            File destDir = new File(destinationDir, resourcePathInJar);
+            if (destDir.exists() && !destDir.isDirectory()) {
+                throw new IOException("The destination path is not a directory: " + destDir);
+            }
+
+            Enumeration<JarEntry> entries = jarFile.entries();
+            while (entries.hasMoreElements()) {
+                JarEntry entry = entries.nextElement();
+
+                if (entry.getName()
+                         .startsWith(resourcePathInJar)) {
+                    String entryPath = entry.getName();
+                    String relativeEntryPath = entryPath.substring(resourcePathInJar.length());
+
+                    File entryDestination = new File(destinationDir, relativeEntryPath);
+
+                    if (entry.isDirectory()) {
+                        if (!entryDestination.exists()) {
+                            entryDestination.mkdirs();
+                        }
+                    } else {
+                        File parentDir = entryDestination.getParentFile();
+                        if (parentDir != null && !parentDir.exists()) {
+                            parentDir.mkdirs(); // Ensure parent directories exist
+                        }
+
+                        try (InputStream is = jarFile.getInputStream(entry); OutputStream os = new FileOutputStream(entryDestination)) {
+                            byte[] buffer = new byte[4096];
+                            int bytesRead;
+                            while ((bytesRead = is.read(buffer)) != -1) {
+                                os.write(buffer, 0, bytesRead);
+                            }
+                        }
+                    }
+                }
+            }
+        }
     }
 
     private void replacePlaceholderInFile(File file, Map<String, String> placeholders) {
